Text classification
This is a work of university project in Data mining.
We a have a large datasets of documents and our goal is to predict in which catecory belong each document. The number of categories are 5 (film,football,politics,technology,business). We use and explore some of machine learning algorithms and visualization tools.
This projects has implemented in Python and has used the SkikitLearn library.