FeatureSelection
This is a work of university project in Data mining.
We a have 1000 users-clients of a bank and our goal is to clasify them if they are good or bad clients according to bank requirements. We use some visualization tools in reports and 3 machine learning algorithms for prediction. In the preprocessing stage we use feature selection for finding the most useful attributes for best gain information.